abstracts main content into its own component
