+++ packages/manager/src/App.tsx
@@ -12,14 +12,7 @@ import { shim } from 'promise.prototype.finally';
 import { path, pathOr } from 'ramda';
 import * as React from 'react';
 import { connect, MapDispatchToProps } from 'react-redux';
-import {
-  Redirect,
-  Route,
-  RouteComponentProps,
-  RouteProps,
-  Switch,
-  withRouter
-} from 'react-router-dom';
+import { RouteComponentProps } from 'react-router-dom';
 import { Action, compose } from 'redux';
 import { ThunkDispatch } from 'redux-thunk';
 import { Subscription } from 'rxjs/Subscription';
@@ -29,14 +22,10 @@ import {
   withStyles,
   WithStyles
 } from 'src/components/core/styles';
-import DefaultLoader from 'src/components/DefaultLoader';
 import {
   DocumentTitleSegment,
   withDocumentTitleProvider
 } from 'src/components/DocumentTitle';
-import Grid from 'src/components/Grid';
-import LandingLoading from 'src/components/LandingLoading';
-import NotFound from 'src/components/NotFound';
 import SideMenu from 'src/components/SideMenu';
 import withFeatureFlagProvider from 'src/containers/withFeatureFlagProvider.container';
 import { events$ } from 'src/events';
@@ -59,91 +48,15 @@ import {
 } from './utilities/accountCapabilities';
 
 import DataLoadedListener from 'src/components/DataLoadedListener';
-import ErrorState from 'src/components/ErrorState';
 import { handleLoadingDone } from 'src/store/initialLoad/initialLoad.actions';
 import { addNotificationsToLinodes } from 'src/store/linodes/linodes.actions';
 import { formatDate } from 'src/utilities/formatDate';
 
 import IdentifyUser from './IdentifyUser';
 
+import MainContent from './MainContent';
+
 shim(); // allows for .finally() usage
-
-const Account = DefaultLoader({
-  loader: () => import('src/features/Account')
-});
-
-const LinodesRoutes = DefaultLoader({
-  loader: () => import('src/features/linodes')
-  // loading: () => <div>loading...</div>
-});
-
-const Volumes = DefaultLoader({
-  loader: () => import('src/features/Volumes')
-});
-
-const Domains = DefaultLoader({
-  loader: () => import('src/features/Domains')
-});
-
-const Images = DefaultLoader({
-  loader: () => import('src/features/Images')
-});
-
-const Kubernetes = DefaultLoader({
-  loader: () => import('src/features/Kubernetes')
-});
-
-const ObjectStorage = DefaultLoader({
-  loader: () => import('src/features/ObjectStorage')
-});
-
-const Profile = DefaultLoader({
-  loader: () => import('src/features/Profile')
-});
-
-const NodeBalancers = DefaultLoader({
-  loader: () => import('src/features/NodeBalancers')
-});
-
-const StackScripts = DefaultLoader({
-  loader: () => import('src/features/StackScripts')
-});
-
-const SupportTickets = DefaultLoader({
-  loader: () => import('src/features/Support/SupportTickets')
-});
-
-const SupportTicketDetail = DefaultLoader({
-  loader: () => import('src/features/Support/SupportTicketDetail')
-});
-
-const Longview = DefaultLoader({
-  loader: () => import('src/features/Longview')
-});
-
-const Managed = DefaultLoader({
-  loader: () => import('src/features/Managed')
-});
-
-const Dashboard = DefaultLoader({
-  loader: () => import('src/features/Dashboard')
-});
-
-const Help = DefaultLoader({
-  loader: () => import('src/features/Help')
-});
-
-const SupportSearchLanding = DefaultLoader({
-  loader: () => import('src/features/Help/SupportSearchLanding')
-});
-
-const SearchLanding = DefaultLoader({
-  loader: () => import('src/features/Search')
-});
-
-const EventsLanding = DefaultLoader({
-  loader: () => import('src/features/Events/EventsLanding')
-});
 
 type ClassNames =
   | 'hidden'
@@ -172,30 +85,6 @@ const styles = (theme: Theme) =>
         marginLeft: theme.spacing(22) + 99 // 275
       }
     },
-    wrapper: {
-      padding: theme.spacing(3),
-      transition: theme.transitions.create('opacity'),
-      [theme.breakpoints.down('sm')]: {
-        paddingTop: theme.spacing(2),
-        paddingLeft: theme.spacing(2),
-        paddingRight: theme.spacing(2)
-      }
-    },
-    grid: {
-      [theme.breakpoints.up('lg')]: {
-        height: '100%'
-      }
-    },
-    switchWrapper: {
-      flex: 1,
-      maxWidth: '100%',
-      position: 'relative',
-      '&.mlMain': {
-        [theme.breakpoints.up('lg')]: {
-          maxWidth: '78.8%'
-        }
-      }
-    },
     hidden: {
       display: 'none',
       overflow: 'hidden'
@@ -205,7 +94,8 @@ const styles = (theme: Theme) =>
 interface Props {
   toggleTheme: () => void;
   toggleSpacing: () => void;
-  location: RouteProps['location'];
+  location: RouteComponentProps['location'];
+  history: RouteComponentProps['history'];
 }
 
 interface State {
@@ -403,8 +293,6 @@ export class App extends React.Component<CombinedProps, State> {
       return null;
     }
 
-    const isKubernetesEnabled = _isKubernetesEnabled(accountCapabilities);
-
     return (
       <React.Fragment>
         <a href="#main-content" className="visually-hidden">
@@ -469,58 +357,11 @@ export class App extends React.Component<CombinedProps, State> {
                 isLoggedInAsCustomer={this.props.isLoggedInAsCustomer}
                 username={this.props.username}
               />
-              <div className={classes.wrapper} id="main-content">
-                <Grid container spacing={0} className={classes.grid}>
-                  <Grid item className={classes.switchWrapper}>
-                    <Switch>
-                      <Route path="/linodes" component={LinodesRoutes} />
-                      <Route path="/volumes" component={Volumes} exact strict />
-                      <Redirect path="/volumes*" to="/volumes" />
-                      <Route path="/nodebalancers" component={NodeBalancers} />
-                      <Route path="/domains" component={Domains} />
-                      <Route path="/managed" component={Managed} />
-                      <Route exact path="/longview" component={Longview} />
-                      <Route exact strict path="/images" component={Images} />
-                      <Redirect path="/images*" to="/images" />
-                      <Route path="/stackscripts" component={StackScripts} />
-                      {getObjectStorageRoute(
-                        accountLoading,
-                        accountCapabilities,
-                        accountError
-                      )}
-                      {isKubernetesEnabled && (
-                        <Route path="/kubernetes" component={Kubernetes} />
-                      )}
-                      <Route path="/account" component={Account} />
-                      <Route
-                        exact
-                        strict
-                        path="/support/tickets"
-                        component={SupportTickets}
-                      />
-                      <Route
-                        path="/support/tickets/:ticketId"
-                        component={SupportTicketDetail}
-                        exact
-                        strict
-                      />
-                      <Route path="/profile" component={Profile} />
-                      <Route exact path="/support" component={Help} />
-                      <Route
-                        exact
-                        strict
-                        path="/support/search/"
-                        component={SupportSearchLanding}
-                      />
-                      <Route path="/dashboard" component={Dashboard} />
-                      <Route path="/search" component={SearchLanding} />
-                      <Route path="/events" component={EventsLanding} />
-                      <Redirect exact from="/" to="/dashboard" />
-                      <Route component={NotFound} />
-                    </Switch>
-                  </Grid>
-                </Grid>
-              </div>
+              <MainContent
+                accountCapabilities={accountCapabilities}
+                accountError={accountError}
+                accountLoading={accountLoading}
+              />
             </main>
             <Footer />
             <WelcomeBanner
@@ -539,35 +380,6 @@ export class App extends React.Component<CombinedProps, State> {
     );
   }
 }
-
-// Render the correct <Route /> component for Object Storage,
-// depending on whether /account is loading or has errors, and
-// whether or not the feature is enabled for this account.
-const getObjectStorageRoute = (
-  accountLoading: boolean,
-  accountCapabilities: AccountCapability[],
-  accountError?: Error | Linode.ApiFieldError[]
-) => {
-  let component;
-
-  if (accountLoading) {
-    component = () => <LandingLoading delayInMS={1000} />;
-  } else if (accountError) {
-    component = () => (
-      <ErrorState errorText="An error has occurred. Please reload and try again." />
-    );
-  } else if (isObjectStorageEnabled(accountCapabilities)) {
-    component = ObjectStorage;
-  }
-
-  // If Object Storage is not enabled for this account, return `null`,
-  // which will appear as a 404
-  if (!component) {
-    return null;
-  }
-
-  return <Route path="/object-storage" component={component} />;
-};
 
 interface DispatchProps {
   addNotificationsToLinodes: (
@@ -685,8 +497,7 @@ export default compose(
   styled,
   withDocumentTitleProvider,
   withSnackbar,
-  withFeatureFlagProvider,
-  withRouter
+  withFeatureFlagProvider
 )(App);
 
 export const hasOauthError = (
