+++ packages/manager/src/components/AccountActivation/AccountActivationError.tsx
@@ -1,10 +1,18 @@
 import * as React from 'react';
 import { compose } from 'recompose';
-import withGlobalErrors, { Props, ReduxState } from 'src/containers/globalErrors.container'
+import withGlobalErrors, { Props } from 'src/containers/globalErrors.container';
 
 type CombinedProps = Props;
 
 const AccountActivationError: React.FC<CombinedProps> = props => {
+  React.useEffect(() => {
+    /** set an account_unactivated error if one hasn't already been set */
+    if (!props.globalErrors.account_unactivated) {
+      props.setErrors({
+        account_unactivated: true
+      });
+    }
+  }, [props.globalErrors]);
 
   return (
     <React.Fragment>
@@ -13,13 +21,7 @@ const AccountActivationError: React.FC<CombinedProps> = props => {
   );
 };
 
-interface S {
-  hello: ReduxState
-}
-
 export default compose<CombinedProps, {}>(
   React.memo,
-  withGlobalErrors<S, {}>((errors => ({
-    hello: errors
-  })))
+  withGlobalErrors()
 )(AccountActivationError);
